[FIX] Fix ZED UI don't start when without internet connection

+++ boot.sh
@@ -2,12 +2,20 @@
 
 if [ -e ".branch" ]
 then
-   git checkout "$(<.branch)"
-   rm .branch
+   echo -e "GET http://google.com HTTP/1.0\n\n" | nc google.com 80 > /dev/null 2>&1
+   if [ $? -eq 0 ]; then
+      git checkout "$(<.branch)"
+      rm .branch
+   fi
 fi
-git stash
-git pull
-git stash pop
+
+echo -e "GET http://google.com HTTP/1.0\n\n" | nc google.com 80 > /dev/null 2>&1
+
+if [ $? -eq 0 ]; then
+   git stash
+   git pull
+   git stash pop
+fi
 cd Backend
 ./startBackend.sh&
 cd ..
